Added reversible field line tracer
The standard FieldTracer seems to work well on some
SciPy versions, but not so well on others. A possible cause is
that the time integrator is not reversible in y.
When integrating over many periods, this can lead to a gradual drift.

FieldTracerReversible uses an implicit trapezoidal (Crank-Nicholson)
algorithm, which is symmetric in y. This means that if run backwards
it should reproduce the same points.

Added an example which makes a Poincare plot