Fixed boundaries for RectangularPoloidalGrids
Boundary code in FCI had not been changed to use the new
fields in Zoidberg output. This is now fixed, so boundaries
should be added to the map correctly if they are marked by Zoidberg.

The RectangularPoloidalGrid didn't properly put the domain
boundaries between grid cells, but on cells instead.
This is now fixed so boundary locations are consistent with BOUT++.

Now converges at 2nd order with periodic or non-periodic Y.
Test is set to non-periodic, but only two resolutions so the test
hopefully isn't too slow.