Zoidberg field tracer now handles large array sizes
Previously all points were evolved together. This is much faster than
evolving one by one, but if large grids were needed (> approx 100 x 100)
then odeint would fail with memory errors.

This patch breaks large grids into pieces, of around 1000 points,
and integrates them separately.

In principle this could be parallelised using multiprocessing.map,
but this needs to be able to pickle code and so struggles with
closures. In a quick test I could get it to work in Python 3, but not
Python 2. For now just serial.