Improve stiff-ness of radial coordinates
The poloidal coordinate needs to be more curved. Treating x and z the
same gives big cells close to corners. It generally pulls the grid
closer to the centre, which is proably not what we want.
the factor should probably be configuarable. 0 means to treat x and z
the same, i.e. the old behaviour. eps->-1 means that x is fixed and
nothing is changed, and we get something closer to grid_annulus