Allow distributing the points on the inner surface more smoothly
This commits adds the option `inner_maxmode`, which can be used in combination
with `inner_ort` to distribute the points on the inner boundary more smoothly.
Unlike `maxmode_inner`, this does not sets a maximum ratio, but ensures that
the points are distributed locally smoothly, by removing high frequency modes
in the distrbution.

This is not enabled by default, as the value of inner_maxmode depends on
the the specific mesh. But it seems that for simple cases, low values
around 5 seem to be good.